+++ scripts/infraPrep.sh
@@ -1,5 +1,5 @@
 #!/bin/bash
-echo $(date) " - Starting Master Prep Script"
+echo $(date) " - Starting Node Prep Script"
 
 export USERNAME_ORG=$1
 export PASSWORD_ACT_KEY="$2"
@@ -51,7 +51,7 @@ else
     else
         echo "Incorrect Pool ID or no entitlements available"
         exit 4
-   fi
+    fi
 fi
 
 # Disable all repositories and enable only the required ones
@@ -65,14 +65,14 @@ subscription-manager repos \
     --enable="rhel-7-server-ose-3.11-rpms" \
     --enable="rhel-7-server-ansible-2.6-rpms" \
     --enable="rhel-7-fast-datapath-rpms" \
-    --enable="rh-gluster-3-client-for-rhel-7-server-rpms" \
-    --enable="rhel-7-server-optional-rpms"
+    --enable="rh-gluster-3-client-for-rhel-7-server-rpms"
 
 # Install base packages and update system to latest packages
 echo $(date) " - Install base packages and update system to latest packages"
 
-yum -y install wget git net-tools bind-utils iptables-services bridge-utils bash-completion httpd-tools kexec-tools sos psacct ansible
+yum -y install wget git net-tools bind-utils iptables-services bridge-utils bash-completion kexec-tools sos psacct
 yum -y install cloud-utils-growpart.noarch
+yum -y install ansible
 yum -y update glusterfs-fuse
 yum -y update --exclude=WALinuxAgent
 echo $(date) " - Base package insallation and updates complete"
@@ -89,17 +89,9 @@ part_number=${name#*${rootdrivename}}
 growpart $rootdrive $part_number -u on
 xfs_growfs $rootdev
 
-if [ $? -eq 0 ]
-then
-    echo "Root partition expanded"
-else
-    echo "Root partition failed to expand"
-    exit 6
-fi
-
 # Install Docker
 echo $(date) " - Installing Docker"
-yum -y install docker 
+yum -y install docker
 
 # Update docker config for insecure registry
 echo "
@@ -108,9 +100,9 @@ OPTIONS=\"\$OPTIONS --insecure-registry 172.30.0.0/16\"
 " >> /etc/sysconfig/docker
 
 # Create thin pool logical volume for Docker
-echo $(date) " - Creating thin pool logical volume for Docker and staring service"
+echo $(date) " - Creating thin pool logical volume for Docker and starting service"
 
-DOCKERVG=$( parted -m /dev/sda print all 2>/dev/null | grep unknown | grep /dev/sd | cut -d':' -f1 )
+DOCKERVG=$( parted -m /dev/sda print all 2>/dev/null | grep unknown | grep /dev/sd | cut -d':' -f1 | head -n1 )
 
 echo "
 # Adding OpenShift data disk for docker
@@ -122,10 +114,10 @@ VG=docker-vg
 docker-storage-setup
 if [ $? -eq 0 ]
 then
-   echo "Docker thin pool logical volume created successfully"
+    echo "Docker thin pool logical volume created successfully"
 else
-   echo "Error creating logical volume for Docker"
-   exit 5
+    echo "Error creating logical volume for Docker"
+    exit 5
 fi
 
 # Enable and start Docker services
@@ -134,3 +126,4 @@ systemctl enable docker
 systemctl start docker
 
 echo $(date) " - Script Complete"
+