+++ scripts/nodePrep.sh
@@ -1,9 +1,9 @@
 #!/bin/bash
 echo $(date) " - Starting Infra / Node Prep Script"
 
-USERNAME_ORG=$1
-PASSWORD_ACT_KEY="$2"
-POOL_ID=$3
+export USERNAME_ORG=$1
+export PASSWORD_ACT_KEY="$2"
+export POOL_ID=$3
 
 # Remove RHUI
 
@@ -13,32 +13,45 @@ sleep 10
 # Register Host with Cloud Access Subscription
 echo $(date) " - Register host with Cloud Access Subscription"
 
-subscription-manager register --username="$USERNAME_ORG" --password="$PASSWORD_ACT_KEY" || subscription-manager register --activationkey="$PASSWORD_ACT_KEY" --org="$USERNAME_ORG"
+subscription-manager register --force --username="$USERNAME_ORG" --password="$PASSWORD_ACT_KEY" || subscription-manager register --force --activationkey="$PASSWORD_ACT_KEY" --org="$USERNAME_ORG"
+RETCODE=$?
 
-if [ $? -eq 0 ]
+if [ $RETCODE -eq 0 ]
 then
-   echo "Subscribed successfully"
-elif [ $? -eq 64 ]
-   then
-       echo "This system is already registered."
+    echo "Subscribed successfully"
+elif [ $RETCODE -eq 64 ]
+then
+    echo "This system is already registered."
 else
-   echo "Incorrect Username / Password or Organization ID / Activation Key specified"
-   exit 3
+    sleep 5
+	subscription-manager register --force --username="$USERNAME_ORG" --password="$PASSWORD_ACT_KEY" || subscription-manager register --force --activationkey="$PASSWORD_ACT_KEY" --org="$USERNAME_ORG"
+	RETCODE2=$?
+	if [ $RETCODE2 -eq 0 ]
+	then
+		echo "Subscribed successfully"
+	elif [ $RETCODE2 -eq 64 ]
+	then
+		echo "This system is already registered."
+	else
+		echo "Incorrect Username / Password or Organization ID / Activation Key specified. Unregistering system from RHSM"
+		subscription-manager unregister
+		exit 3
+	fi
 fi
 
 subscription-manager attach --pool=$POOL_ID > attach.log
 if [ $? -eq 0 ]
 then
-   echo "Pool attached successfully"
+    echo "Pool attached successfully"
 else
-   evaluate=$( cut -f 2-5 -d ' ' attach.log )
-   if [[ $evaluate == "unit has already had" ]]
-      then
-         echo "Pool $POOL_ID was already attached and was not attached again."
-	  else
-         echo "Incorrect Pool ID or no entitlements available"
-         exit 4
-   fi
+    grep attached attach.log
+    if [ $? -eq 0 ]
+    then
+        echo "Pool $POOL_ID was already attached and was not attached again."
+    else
+        echo "Incorrect Pool ID or no entitlements available"
+        exit 4
+    fi
 fi
 
 # Disable all repositories and enable only the required ones
@@ -49,19 +62,20 @@ subscription-manager repos --disable="*"
 subscription-manager repos \
     --enable="rhel-7-server-rpms" \
     --enable="rhel-7-server-extras-rpms" \
-    --enable="rhel-7-server-ose-3.9-rpms" \
-    --enable="rhel-7-server-ansible-2.4-rpms" \
-    --enable="rhel-7-fast-datapath-rpms"
+    --enable="rhel-7-server-ose-3.11-rpms" \
+    --enable="rhel-7-server-ansible-2.6-rpms" \
+    --enable="rhel-7-fast-datapath-rpms" \
+    --enable="rh-gluster-3-client-for-rhel-7-server-rpms" \
+    --enable="rhel-7-server-optional-rpms"
 
 # Install base packages and update system to latest packages
 echo $(date) " - Install base packages and update system to latest packages"
 
-yum -y install wget git net-tools bind-utils iptables-services bridge-utils bash-completion kexec-tools sos psacct
+yum -y install wget git net-tools bind-utils iptables-services bridge-utils bash-completion kexec-tools sos psacct ansible
 yum -y install cloud-utils-growpart.noarch
+yum -y update glusterfs-fuse
 yum -y update --exclude=WALinuxAgent
-yum -y install atomic-openshift-excluder atomic-openshift-docker-excluder
-
-atomic-openshift-excluder unexclude
+echo $(date) " - Base package insallation and updates complete"
 
 # Grow Root File System
 echo $(date) " - Grow Root FS"
@@ -75,26 +89,43 @@ part_number=${name#*${rootdrivename}}
 growpart $rootdrive $part_number -u on
 xfs_growfs $rootdev
 
+if [ $? -eq 0 ]
+then
+    echo "Root partition expanded"
+else
+    echo "Root partition failed to expand"
+    exit 6
+fi
+
 # Install Docker
 echo $(date) " - Installing Docker"
 yum -y install docker
 
-sed -i -e "s#^OPTIONS='--selinux-enabled'#OPTIONS='--selinux-enabled --insecure-registry 172.30.0.0/16'#" /etc/sysconfig/docker
+# Update docker config for insecure registry
+echo "
+# Adding insecure-registry option required by OpenShift
+OPTIONS=\"\$OPTIONS --insecure-registry 172.30.0.0/16\"
+" >> /etc/sysconfig/docker
 
 # Create thin pool logical volume for Docker
 echo $(date) " - Creating thin pool logical volume for Docker and staring service"
 
-DOCKERVG=$( parted -m /dev/sda print all 2>/dev/null | grep unknown | grep /dev/sd | cut -d':' -f1 )
+DOCKERVG=$( parted -m /dev/sda print all 2>/dev/null | grep unknown | grep /dev/sd | cut -d':' -f1 | head -n1 )
 
-echo "DEVS=${DOCKERVG}" >> /etc/sysconfig/docker-storage-setup
-echo "VG=docker-vg" >> /etc/sysconfig/docker-storage-setup
+echo "
+# Adding OpenShift data disk for docker
+DEVS=${DOCKERVG}
+VG=docker-vg
+" >> /etc/sysconfig/docker-storage-setup
+
+# Running setup for docker storage
 docker-storage-setup
 if [ $? -eq 0 ]
 then
-   echo "Docker thin pool logical volume created successfully"
+    echo "Docker thin pool logical volume created successfully"
 else
-   echo "Error creating logical volume for Docker"
-   exit 5
+    echo "Error creating logical volume for Docker"
+    exit 5
 fi
 
 # Enable and start Docker services